+++ CHANGELOG.md
@@ -29,3 +29,7 @@
   
 ## 0.0.8
 * Added the ability to pass gas multiplier values to the SmartWalletFactory class, enabling customized gas.
+
+## 0.0.9
+* Introduced the ability to pass a custom `http.Client` instance to the `SmartWalletFactory`, `JsonRPCProvider`, and `BundlerProvider` classes.
+  - This can allow to pass SentryHttpClient to catch http rpc errors